added class TmcMotionControl #93 - added inherited class TmcMotionControlStepDir - added inherited class TmcMotionControlVActual
