+++ demo/demo_script_03_basic_movement.py
@@ -26,14 +26,14 @@ print("---")
 # use your pins for pin_en, pin_step, pin_dir here
 #-----------------------------------------------------------------------
 if BOARD == Board.RASPBERRY_PI:
-    tmc = Tmc2209(21, 16, 20, TmcUart("/dev/serial0"), loglevel=Loglevel.DEBUG)
+    tmc = Tmc2209(TmcMotionControlStepDir(16, 20), 21, TmcComUart("/dev/serial0"), loglevel=Loglevel.DEBUG)
 elif BOARD == Board.RASPBERRY_PI5:
-    tmc = Tmc2209(21, 16, 20, TmcUart("/dev/ttyAMA0"), loglevel=Loglevel.DEBUG)
+    tmc = Tmc2209(TmcMotionControlStepDir(16, 20), 21, TmcComUart("/dev/ttyAMA0"), loglevel=Loglevel.DEBUG)
 elif BOARD == Board.NVIDIA_JETSON:
-    tmc = Tmc2209(13, 6, 5, TmcUart("/dev/ttyTHS1"), loglevel=Loglevel.DEBUG)
+    tmc = Tmc2209(TmcMotionControlStepDir(6, 5), 13, TmcComUart("/dev/ttyTHS1"), loglevel=Loglevel.DEBUG)
 else:
     # just in case
-    tmc = Tmc2209(21, 16, 20, TmcUart("/dev/serial0"), loglevel=Loglevel.DEBUG)
+    tmc = Tmc2209(TmcMotionControlStepDir(16, 20), 21, TmcComUart("/dev/serial0"), loglevel=Loglevel.DEBUG)
