+++ src/tmc_driver/com/_tmc_com_uart.py
@@ -1,33 +1,23 @@
 #pylint: disable=import-error
 #pylint: disable=broad-exception-caught
 """
-TmcUart stepper driver uart module
+TmcComUart stepper driver uart module
 """
 
-import time
-import struct
-from typing import List
 import serial
-from .reg._tmc_220x_reg_addr import TmcRegAddr
-from .reg._tmc_gstat import GStat
-from ._tmc_logger import TmcLogger, Loglevel
+from ._tmc_com import *
 
 
-class TmcUart:
-    """TmcUart
+class TmcComUart(TmcCom):
+    """TmcComUart
 
     this class is used to communicate with the TMC via UART
     it can be used to change the settings of the TMC.
     like the current or the microsteppingmode
     """
-    _tmc_logger:TmcLogger = None
 
-    mtr_id:int = 0
     ser:serial.Serial = None
-    r_frame:List[int] = [0x55, 0, 0, 0  ]
-    w_frame:List[int] = [0x55, 0, 0, 0 , 0, 0, 0, 0 ]
-    communication_pause:int = 0
-    error_handler_running:bool = False
+
 
     @property
     def tmc_logger(self) -> TmcLogger:
@@ -88,34 +78,10 @@ class TmcUart:
         self.ser.reset_input_buffer()
 
 
-
     def __del__(self):
         """destructor"""
         if self.ser is not None and isinstance(self.ser, serial.Serial):
             self.ser.close()
-
-
-
-    def compute_crc8_atm(self, datagram, initial_value=0):
-        """this function calculates the crc8 parity bit
-
-        Args:
-            datagram (list): datagram
-            initial_value (int): initial value (Default value = 0)
-        """
-        crc = initial_value
-        # Iterate bytes in data
-        for byte in datagram:
-            # Iterate bits in byte
-            for _ in range(0, 8):
-                if (crc >> 7) ^ (byte & 0x01):
-                    crc = ((crc << 1) ^ 0x07) & 0xFF
-                else:
-                    crc = (crc << 1) & 0xFF
-                # Shift to next bit
-                byte = byte >> 1
-        return crc
-
 
 
     def read_reg(self, register:TmcRegAddr):
@@ -153,7 +119,6 @@ class TmcUart:
         return rtn
 
 
-
     def read_int(self, register:TmcRegAddr, tries:int = 10):
         """this function tries to read the registry of the TMC 10 times
         if a valid answer is returned, this function returns it as an integer
@@ -189,7 +154,6 @@ class TmcUart:
 
         val = struct.unpack(">i",rtn_data)[0]
         return val
-
 
 
     def write_reg(self, register:TmcRegAddr, val:int):
@@ -230,7 +194,6 @@ class TmcUart:
         return True
 
 
-
     def write_reg_check(self, register:TmcRegAddr, val:int, tries:int=10):
         """this function als writes a value to the register of the TMC
         but it also checks if the writing process was successfully by checking
@@ -264,14 +227,12 @@ class TmcUart:
                 return -1
 
 
-
     def flush_serial_buffer(self):
         """this function clear the communication buffers of the Raspberry Pi"""
         if self.ser is None:
             return
         self.ser.reset_output_buffer()
         self.ser.reset_input_buffer()
-
 
 
     def handle_error(self):
@@ -298,7 +259,6 @@ class TmcUart:
                       The driver is disabled in this case""", Loglevel.DEBUG)
         self._tmc_logger.log("EXITING!", Loglevel.INFO)
         raise SystemExit
-
 
 
     def test_com(self, register:TmcRegAddr):
