+++ tests/test_tmc_move.py
@@ -9,6 +9,7 @@
 test for _tmc_move.py
 """
 
+import time
 import unittest
 from src.tmc_driver.tmc_2209 import *
 
@@ -17,7 +18,7 @@ class TestTMCMove(unittest.TestCase):
 
     def setUp(self):
         """setUp"""
-        self.tmc = Tmc2209(21, 16, 20)
+        self.tmc = Tmc2209(TmcMotionControlStepDir(16, 20), 21)
 
         # these values are normally set by reading the driver
         self.tmc.mres = 2
@@ -34,38 +35,38 @@ class TestTMCMove(unittest.TestCase):
         """test_run_to_position_steps"""
 
         self.tmc.run_to_position_steps(400, MovementAbsRel.RELATIVE)
-        pos = self.tmc.current_pos
+        pos = self.tmc.tmc_mc.current_pos
         self.assertEqual(pos, 400, f"actual position: {pos}, expected position: 400")
 
         self.tmc.run_to_position_steps(-200, MovementAbsRel.RELATIVE)
-        pos = self.tmc.current_pos
+        pos = self.tmc.tmc_mc.current_pos
         self.assertEqual(pos, 200, f"actual position: {pos}, expected position: 200")
 
         self.tmc.run_to_position_steps(400)
-        pos = self.tmc.current_pos
+        pos = self.tmc.tmc_mc.current_pos
         self.assertEqual(pos, 400, f"actual position: {pos}, expected position: 400")
 
     def test_run_to_position_steps_threaded(self):
         """test_run_to_position_steps_threaded"""
-        self.tmc.run_to_position_steps_threaded(400, MovementAbsRel.RELATIVE)
-        self.tmc.wait_for_movement_finished_threaded()
-        pos = self.tmc.current_pos
+        self.tmc.tmc_mc.run_to_position_steps_threaded(400, MovementAbsRel.RELATIVE)
+        self.tmc.tmc_mc.wait_for_movement_finished_threaded()
+        pos = self.tmc.tmc_mc.current_pos
         self.assertEqual(pos, 400, f"actual position: {pos}, expected position: 400")
 
-        self.tmc.run_to_position_steps_threaded(-200, MovementAbsRel.RELATIVE)
-        self.tmc.wait_for_movement_finished_threaded()
-        pos = self.tmc.current_pos
+        self.tmc.tmc_mc.run_to_position_steps_threaded(-200, MovementAbsRel.RELATIVE)
+        self.tmc.tmc_mc.wait_for_movement_finished_threaded()
+        pos = self.tmc.tmc_mc.current_pos
         self.assertEqual(pos, 200, f"actual position: {pos}, expected position: 200")
 
-        self.tmc.run_to_position_steps_threaded(400)
-        self.tmc.wait_for_movement_finished_threaded()
-        pos = self.tmc.current_pos
+        self.tmc.tmc_mc.run_to_position_steps_threaded(400)
+        self.tmc.tmc_mc.wait_for_movement_finished_threaded()
+        pos = self.tmc.tmc_mc.current_pos
         self.assertEqual(pos, 400, f"actual position: {pos}, expected position: 400")
 
-        self.tmc.run_to_position_steps_threaded(800)
+        self.tmc.tmc_mc.run_to_position_steps_threaded(800)
         time.sleep(0.05)
-        self.tmc.stop()
-        pos = self.tmc.current_pos
+        self.tmc.tmc_mc.stop()
+        pos = self.tmc.tmc_mc.current_pos
         print(f"motorposition: {pos}")
         self.assertTrue(400 < pos < 800, f"actual position: {pos}, expected position: 400 < pos < 800")
